[Feature][Olana] Integrated my-items into tabs
